Commit 22thNov
The class Zombie was created, by inheritance from the sprite class of the pygame librery. Methods was added to make the zombies chase the hero.

The hero class was change to inherit also from the sprite class.

In the main, the creation of zombies with random time and position has been added, using sprite groups.

Next steps I'm going to do: add collisions between zombies and hero and its effect on the hero lifes.