Collision and lives conter added
Collision between the zombies and the hero was added. Lives counter was added and displayed in the screen. You are going to have to install the 8 bits madness font(it's added in a folder, just double click on the tff file). I'll try to use the path to do it independent from the computer (if you come up with this solution you can also modify it)